==> notebooks/1_wiki_article_page_info_retrival.ipynb <==
# Article Page Info MediaWiki API Example
This notebook contains the code required to retreive page information from the MediaWiki API. The code is written in Python and uses the `requests` library to make the API calls. The code is designed to be run in a Jupyter notebook, but could be adapted to run in a standalone Python script.

## License
Some part of the code are used form the examples developed by Dr. David W. McDonald for use in DATA 512, a course in the UW MS Data Science degree program. This code is provided under the [Creative Commons](https://creativecommons.org) [CC-BY license](https://creativecommons.org/licenses/by/4.0/). Revision 1.2 - September 16, 2024



List of required packages:
1. json
2. time
3. urllib
4. dotenv
5. requests: The 'requests' module is not a standard Python module. You will need to install this with pip/pip3 if you do not already have it.
6. pandas

# 
# These are standard python modules
import json, time, urllib.parse
#
import pandas as pd
# The 'requests' module is not a standard Python module. You will need to install this with pip/pip3 if you do not already have it
import requests

First we are retriving the politicians data provided in the assignment.

### Make sure to copy the csv file from the data_files folder and place it in the path of this notebook.

csv_file = pd.read_csv('politicians_by_country_AUG.2024.csv')
print(csv_file)

For each article_name(encoded as name), we want to get the lastrevid which is the revision id of the last edit to the page.

For this, we start by first defining the base url and the parameters required to get the page information.

#########
#
#    CONSTANTS
#

# The basic English Wikipedia API endpoint
API_ENWIKIPEDIA_ENDPOINT = "https://en.wikipedia.org/w/api.php"
API_HEADER_AGENT = 'User-Agent'

# We'll assume that there needs to be some throttling for these requests - we should always be nice to a free data resource
API_LATENCY_ASSUMED = 0.002       # Assuming roughly 2ms latency on the API and network
API_THROTTLE_WAIT = (1.0/100.0)-API_LATENCY_ASSUMED

# When making automated requests we should include something that is unique to the person making the request
# This should include an email - your UW email would be good to put in there
REQUEST_HEADERS = {
    'User-Agent': 'dvabhinav31@gmail.com, University of Washington, MSDS DATA 512 - AUTUMN 2024'
}

# This is just a list of English Wikipedia article titles that we can use for example requests
ARTICLE_TITLES = list(csv_file['name'])
clenaed_article_titles = []
for article in ARTICLE_TITLES:
    clenaed_article_titles.append(article.replace(' ', '_'))

# This is a string of additional page properties that can be returned see the Info documentation for
# what can be included. If you don't want any this can simply be the empty string
PAGEINFO_EXTENDED_PROPERTIES = "talkid|url|watched|watchers"
#PAGEINFO_EXTENDED_PROPERTIES = ""

# This template lists the basic parameters for making this
PAGEINFO_PARAMS_TEMPLATE = {
    "action": "query",
    "format": "json",
    "titles": "",           # to simplify this should be a single page title at a time
    "prop": "info",
    "inprop": PAGEINFO_EXTENDED_PROPERTIES
}


The API request will be made using one procedure. The idea is to make this reusable. The procedure is parameterized, but relies on the constants above for the important parameters. The underlying assumption is that this will be used to request data for a set of article pages. Therefore the parameter most likely to change is the name.



#########
#
#    PROCEDURES/FUNCTIONS
#

def request_pageinfo_per_article(article_title = None, 
                                 endpoint_url = API_ENWIKIPEDIA_ENDPOINT, 
                                 request_template = PAGEINFO_PARAMS_TEMPLATE,
                                 headers = REQUEST_HEADERS):
    
    # article title can be as a parameter to the call or in the request_template
    if article_title:
        request_template['titles'] = article_title

    if not request_template['titles']:
        raise Exception("Must supply an article title to make a pageinfo request.")

    if API_HEADER_AGENT not in headers:
        raise Exception(f"The header data should include a '{API_HEADER_AGENT}' field that contains your UW email address.")

    if 'uwnetid@uw' in headers[API_HEADER_AGENT]:
        raise Exception(f"Use your UW email address in the '{API_HEADER_AGENT}' field.")

    # make the request
    try:
        # we'll wait first, to make sure we don't exceed the limit in the situation where an exception
        # occurs during the request processing - throttling is always a good practice with a free
        # data source like Wikipedia - or any other community sources
        if API_THROTTLE_WAIT > 0.0:
            time.sleep(API_THROTTLE_WAIT)
        response = requests.get(endpoint_url, headers=headers, params=request_template)
        json_response = response.json()
    except Exception as e:
        print(e)
        json_response = None
    return json_response


There is a way to get the information for multiple pages at the same time, by separating the page titles with the vertical bar "|" character. However, this approach has limits. You should probably check the API documentation if you want to do multiple pages in a single request - and limit the number of pages in one request reasonably.

The batch size is set to 50, which is the maximum number of pages that can be requested in a single call. THis significantly reduces the number of API calls required from 7100 to around 142.

We only need the lastreid and the name of the article. We will store this information in a list of lists. The list will be converted to a dataframe and saved as a csv file in the next cell

wiki_information = []
batch_size = 50
for i in range(0, len(clenaed_article_titles), batch_size):
    names = '|'.join(clenaed_article_titles[i:i+batch_size])
    info = request_pageinfo_per_article(names)
    if 'query' in info and 'pages' in info['query']:
        response = info['query']['pages']
        for page_id in response:
            name = response[page_id]['title']
            if 'lastrevid' in response[page_id]:
                wiki_information.append([name,response[page_id]['lastrevid']])
            else:
                wiki_information.append([name,''])

Here, we write the dictionary into a csv for future use in analysis

#groupby the wiki_information based on the revid and name
wiki_information_df = pd.DataFrame(wiki_information, columns = ['name', 'revid'])
wiki_information_df = wiki_information_df.groupby(['revid','name']).size().reset_index(name='count')

# Basic Analysis
There are a total of 7155 articles in the dataset.

I wanted to check if there are any repeated name and revision id pairs. This is important because we are using the revision id as a unique identifier for the page. If there are any duplicates, we need to investigate further.


wiki_information_df[wiki_information_df['count'] > 1]

We do not need these exact duplicates, so we will remove them. We will keep the first instance of the duplicate and remove the rest.

wiki_information_df = wiki_information_df.drop_duplicates(subset=['revid'], keep='first')

#write wiki_information whihc is a list of list with name and response into a csv file
wiki_information_df.to_csv('wiki_information.csv', index = False)

wiki_information_df

len(wiki_information)

len(wiki_information_df)
